# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Firetower",
    "description": "The Firetower control plane: API, scheduling, and worker transports.",
    "contact": {
      "name": "Westlabs LLC"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/agents/{kind}": {
      "put": {
        "tags": [
          "agents"
        ],
        "summary": "Configure how an agent authenticates.",
        "operationId": "configure_agent",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "Agent kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureAgent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "Forget an agent's configuration and any credential with it.",
        "operationId": "forget_agent",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "Agent kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Changed. Every other browser is signed out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rotated"
                }
              }
            }
          },
          "400": {
            "description": "Too short, or the current one is wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Replay. The live feed is the event stream; this is the backfill after a hard\nrefresh, and the fallback when a stream can't be held open.",
        "operationId": "list_events",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Last sequence number seen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "description": "Only this session's events",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hosts": {
      "get": {
        "tags": [
          "hosts"
        ],
        "operationId": "list_hosts",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Host"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Add somewhere for agents to run.",
        "description": "Connecting happens straight away rather than on the next restart, so a\nmistake in an address is a message here instead of a host that silently\nnever works.\n\nA host that doesn't answer is still created, and comes back `Unreachable`\nwith a `diagnosis` saying why and what to run. That is not the same as the\nrequest failing: what someone typed was accepted, and the machine at the\nother end has something to fix. Only what we can rule out from here —\nan empty address, a key that isn't a key, a name already taken — is a 400.",
        "operationId": "create_host",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Host"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hosts/probe": {
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Try a machine without adding it.",
        "description": "Adding used to create the row and then connect, which left a host behind\nwhenever the first attempt failed — and with Firetower's own key, the first\nattempt failing is the ordinary path rather than the exception: a machine has\nnot been told about the key yet.\n\nSo the interface asks this first. Nothing is stored, so a mistyped address\nleaves nothing to delete, and retrying is a button rather than a form to fill\nin again.",
        "operationId": "probe_host",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reached"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hosts/{id}": {
      "delete": {
        "tags": [
          "hosts"
        ],
        "summary": "Forget a host, and take its container with it.",
        "description": "Refuses while sessions are running unless `force`, in which case they are\ntold to end first — an agent that gets to shut down leaves its worktree and\ntmux session behind cleanly, rather than having the floor pulled out.\n\nA container Firetower started is Firetower's to stop. One it merely found\nrunning is not, and start-up says as much when it adopts nothing.",
        "operationId": "delete_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "End running sessions instead of refusing",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "hosts"
        ],
        "summary": "Call it something else.",
        "description": "The name and nothing else: what a host *is* was settled when it was added,\nand pointing it somewhere different is removing it and adding another.",
        "operationId": "rename_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Host"
                }
              }
            }
          },
          "400": {
            "description": "Empty, or already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hosts/{id}/connect": {
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Try a host again now, instead of waiting out the backoff.",
        "description": "The supervisor would get there on its own; this is for the moment just after\nyou have fixed the machine and would rather not wait.",
        "operationId": "connect_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hosts/{id}/drain": {
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Stop sending work here, or start again.",
        "operationId": "drain_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Drain"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/providers/{id}": {
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Sign out: forget the token and stop any authorization still waiting.",
        "operationId": "disconnect_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}/authorize": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Start an authorization and begin waiting for it to be approved.",
        "description": "Returns immediately with the code to show. The waiting happens here rather\nthan in the browser so that closing the tab doesn't abandon it.",
        "operationId": "authorize_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingAuth"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}/client-id": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Register an application to authorize against.",
        "description": "Asked twice on purpose: once in the setup wizard, where it is skippable,\nand again on the connect-a-repository screen at the moment somebody wants\nthe thing it enables. Stored rather than configured, so it takes effect\nwithout a restart.",
        "operationId": "set_client_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Stored"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}/repos": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "What the authorized account can see. This is the picker's data.",
        "operationId": "list_provider_repos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoteRepo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos": {
      "get": {
        "tags": [
          "repos"
        ],
        "operationId": "list_repos",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Repo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "repos"
        ],
        "operationId": "create_repo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRepo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/probe": {
      "post": {
        "tags": [
          "repos"
        ],
        "summary": "Can we reach it, and what is it called?",
        "description": "Answered by a worker, because the worker is what holds the credentials and\nwhat will do the clone. Anything decided here would be a guess about someone\nelse's network.",
        "operationId": "probe_repo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProbeResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Every non-success response, so failures are as typed as everything else.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "For humans and logs. The interface should switch on `code` and write its\nown copy — only it knows the context and what to offer next."
          }
        }
      },
...
      "ConfigureAgent": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/AgentMode"
          },
          "secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "The token from `claude setup-token`, or a metered API key — whichever\nthe mode calls for. Required for both; ignored for an agent that needs\nno credential."
          }
        }
      },
...
      "NewPassword": {
        "type": "object",
        "required": [
          "current",
          "new"
        ],
        "properties": {
          "current": {
            "type": "string",
            "description": "The one in use. Required even when it came from a file: knowing the\nsession token should not be enough to lock the owner out."
          },
          "new": {
            "type": "string"
          }
        }
      },
...
      "Rotated": {
        "type": "object",
        "description": "What the browser that changed a password gets back.",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "A new session, because the old one went with every other. Store this\nand carry on — being signed out is not part of changing a password."
          }
        }
      },
...
    }
  }
}